add flex in profiles
