+++ FrontEnd/src/global.css
@@ -108,4 +108,6 @@
   --main-black-90: #292e32;
   --primary-green-80: #1f9a7c;
   --character-title-85: rgba(0, 0, 0, 0.85);
+
+  --wf-base-white: #FFF;
 }
