+++ FrontEnd/src/pages/CustomThemes/customProfileTooltipThemes.js
@@ -1,0 +1,11 @@
+const tooltipInnerContentStyles = {
+    display: 'flex',
+    borderRadius: '2px',
+    background: 'var(--main-grey-90, #25292C)',
+    color: 'var(--main-white, #FFF)',
+    fontFeatureSettings: 'calt',
+    fontFamily: 'Inter',
+    letterSpacing: '-0.14px',
+};
+
+export default tooltipInnerContentStyles;
