+++ FrontEnd/src/pages/ProfilePage/ProfilePageComponents/ProfileContent.jsx
@@ -2,7 +2,7 @@ import { Tooltip } from 'antd';
 import { PropTypes } from 'prop-types';
 import { Link, NavLink, Route, Routes, useBlocker } from 'react-router-dom';
 import { useEffect, useState } from 'react';
-import { DirtyFormContext } from  '../../../context/DirtyFormContext';
+import { DirtyFormContext } from '../../../context/DirtyFormContext';
 import AdditionalInfo from '../FormComponents/AdditionalInfo';
 import ContactsInfo from '../FormComponents/ContactsInfo';
 import DeleteProfilePage from '../FormComponents/DeleteProfileComponent/DeleteProfilePage';
@@ -15,6 +15,7 @@ import MyModal from '../UI/MyModal/MyModal';
 import WarnUnsavedDataModal from '../FormComponents/WarnUnsavedDataModal';
 import ChangePassword from '../FormComponents/ChangePassword';
 import css from './ProfileContent.module.css';
+import tooltipInnerContentStyles from '../../CustomThemes/customProfileTooltipThemes';
 
 
 const INFOLINKS = [
@@ -59,24 +60,13 @@ const FORM_NAMES = [
 ];
 
 const ProfileContent = (props) => {
-
-    const tooltipInnerContentStyles = {
-            display: 'flex',
-            borderRadius: '2px',
-            background: 'var(--main-grey-90, #25292C)',
-            color: 'var(--main-white, #FFF)',
-            fontFeatureSettings: 'calt',
-            fontFamily: 'Inter',
-            letterSpacing: '-0.14px',
-    };
-
     const [modal, setModal] = useState(false);
     const [formIsDirty, setFormIsDirty] = useState(false);
     const blocker = useBlocker(
         ({ currentLocation, nextLocation }) =>
-          formIsDirty &&
-          currentLocation.pathname !== nextLocation.pathname
-      );
+            formIsDirty &&
+            currentLocation.pathname !== nextLocation.pathname
+    );
 
     useEffect(() => {
         if (blocker.state === 'blocked') {
@@ -90,79 +80,76 @@ const ProfileContent = (props) => {
         setFormIsDirty(false);
         blocker.proceed();
         setModal(false);
-      };
+    };
 
     const cancelNavigation = () => {
         blocker.reset();
         setModal(false);
-      };
+    };
 
-    useEffect (() => {
+    useEffect(() => {
         const onBeforeUnload = (e) => {
             if (formIsDirty) {
-              e.preventDefault();
-              e.returnValue = '';
+                e.preventDefault();
+                e.returnValue = '';
             }
-          };
-          window.addEventListener('beforeunload', onBeforeUnload);
-          return () => {
+        };
+        window.addEventListener('beforeunload', onBeforeUnload);
+        return () => {
             window.removeEventListener('beforeunload', onBeforeUnload);
-          };
-        }, [formIsDirty]);
+        };
+    }, [formIsDirty]);
 
     return (
-        <div className={css['content__container']}>
-            <div className={css['content']}>
-                <div className={css['info-links-profile']}>
-                {INFOLINKS.map((element) => {
-                    const isLinkEnabled =
-                        (props.profile.is_registered && props.profile.is_startup) ||
-                        (props.profile.is_registered && element.title !== 'Стартап') ||
-                        (props.profile.is_startup && element.title !== 'Інформація про товари/ послуги' && element.title !== 'Додаткова інформація');
-                    return isLinkEnabled ? (
-                        <NavLink
-                            className={({ isActive }) => (`${css['infolink']} ${isActive && css['infolink__active']}`)}
-                            to={`/profile${element.link}`}
-                            key={element.title}
-                        >
-                            {element.title}
-                        </NavLink>
-                    ) : (
-                        <Tooltip
-                            key={element.title}
-                            placement="bottomLeft"
-                            title={element.tooltipText}
-                            overlayInnerStyle={tooltipInnerContentStyles}
-                        >
+        <div className={css['profile__main']}>
+            <div className={css['profile__content']}>
+                <div className={css['profile__info-links']}>
+                    {INFOLINKS.map((element) => {
+                        const isLinkEnabled =
+                            (props.profile.is_registered && props.profile.is_startup) ||
+                            (props.profile.is_registered && element.title !== 'Стартап') ||
+                            (props.profile.is_startup && element.title !== 'Інформація про товари/ послуги'
+                                && element.title !== 'Додаткова інформація');
+                        return isLinkEnabled ? (
+                            <NavLink
+                                className={({ isActive }) =>
+                                    (`${css['profile__infolink']} ${isActive && css['profile__infolink-active']}`)}
+                                to={`/profile${element.link}`}
+                                key={element.title}
+                            >
+                                {element.title}
+                            </NavLink>
+                        ) : (
+                            <Tooltip
+                                key={element.title}
+                                placement="bottomLeft"
+                                title={element.tooltipText}
+                                overlayInnerStyle={tooltipInnerContentStyles}
+                            >
                                 <NavLink
-                                    className={`${css['infolink']} ${css['infolink__disabled']}`}
+                                    className={`${css['profile__infolink']} ${css['profile__infolink-disabled']}`}
                                     key={element.title}
                                     disabled
                                 >
                                     {element.title}
                                 </NavLink>
-                        </Tooltip>
-                    );
-                })}
-                    <div className={css['divider']}></div>
+                            </Tooltip>
+                        );
+                    })}
+                    <div className={css['profile__divider']}></div>
                     <Link
                         to="/profile/change-password"
-                        className={`${css['infolink']}`}>
+                        className={`${css['profile__infolink']}`}>
                         Змінити пароль
                     </Link>
                     <Link
                         to="/profile/delete"
-                        className={`${css['infolink']} ${css['delete']}`}>
+                        className={`${css['profile__infolink']} ${css['profile__delete']}`}>
                         Видалити профіль
                     </Link>
                 </div>
                 <DirtyFormContext.Provider value={{ formIsDirty, setFormIsDirty }}>
                     <Routes>
-                        <Route
-                            path="/delete"
-                            element={<DeleteProfilePage
-                                currentFormNameHandler={props.currentFormNameHandler}
-                                curForm={FORM_NAMES[6]} />} />
                         <Route
                             path="/user-info"
                             element={<UserInfo user={props.user}
@@ -199,6 +186,11 @@ const ProfileContent = (props) => {
                                 currentFormNameHandler={props.currentFormNameHandler}
                                 curForm={FORM_NAMES[5]} />} />
                         <Route
+                            path="/delete"
+                            element={<DeleteProfilePage
+                                currentFormNameHandler={props.currentFormNameHandler}
+                                curForm={FORM_NAMES[6]} />} />
+                        <Route
                             path="/change-password"
                             element={<ChangePassword
                                 user={props.user}
@@ -207,16 +199,17 @@ const ProfileContent = (props) => {
                     </Routes>
                 </DirtyFormContext.Provider>
             </div>
+            <div className={css['profile__elastic']} ></div>
+            <div className={css['profile__button']}>
+                {props.formName !== 'Delete' && <ProfileFormButton formName={props.formName} formState={formIsDirty} />}
 
-            {props.formName !== 'Delete' && <ProfileFormButton formName={props.formName} formState={formIsDirty}/>}
-
-            {blocker.state === 'blocked' &&
-                (
-                <MyModal visible={modal}>
-                    <WarnUnsavedDataModal onCancel={cancelNavigation} onConfirm={confirmNavigation} />
-                </MyModal>)
-            }
-
+                {blocker.state === 'blocked' &&
+                    (
+                        <MyModal visible={modal}>
+                            <WarnUnsavedDataModal onCancel={cancelNavigation} onConfirm={confirmNavigation} />
+                        </MyModal>)
+                }
+            </div>
         </div>
     );
 };
